@@ -1,11 +1,10 @@
-// server
-// tests/crypto-utils.test.js
+import test from 'node:test'
+import assert from 'node:assert/strict'
+import { generateSecureSecret } from '../app/crypto-utils.js'
 
-const { generateSecureSecret } = require("../app/crypto-utils")
-
-describe("generateSecureSecret", () => {
-  it("should generate a 64-character hex string", () => {
+test('generateSecureSecret', async (t) => {
+  await t.test('should generate a 64-character hex string', () => {
     const secret = generateSecureSecret()
-    expect(secret).toMatch(/^[0-9a-f]{64}$/)
+    assert.match(secret, /^[0-9a-f]{64}$/)
   })
 })
