@@ -1,291 +1,268 @@
-/* eslint-disable jest/no-conditional-expect */
-// server
-// tests/ssh.test.js
+import ssh2 from 'ssh2'
+import crypto from 'crypto'
+import { test, describe, beforeEach, afterEach } from 'node:test'
+import { strict as assert } from 'assert'
+const { Server } = ssh2
+import SSHConnection from '../app/ssh.js'
 
-const SSH2 = require("ssh2")
-const SSHConnection = require("../app/ssh")
-const { SSHConnectionError } = require("../app/errors")
-const { DEFAULTS } = require("../app/constants")
-
-jest.mock("ssh2")
-jest.mock("../app/logger", () => ({
-  createNamespacedDebug: jest.fn(() => jest.fn()),
-  logError: jest.fn()
-}))
-jest.mock("../app/utils", () => ({
-  maskSensitiveData: jest.fn((data) => data)
-}))
-jest.mock("../app/errors", () => ({
-  SSHConnectionError: jest.fn(function (message) {
-    this.message = message
-  }),
-  handleError: jest.fn()
-}))
-
-describe("SSHConnection", () => {
+describe('SSHConnection', () => {
+  let sshServer
   let sshConnection
-  let mockConfig
-  let mockSSH2Client
-  let registeredEventHandlers
+  const TEST_PORT = 2222
+  const TEST_CREDENTIALS = {
+    username: 'testuser',
+    password: 'testpass',
+  }
+
+  const { privateKey } = crypto.generateKeyPairSync('rsa', {
+    modulusLength: 2048,
+    publicKeyEncoding: {
+      type: 'spki',
+      format: 'pem',
+    },
+    privateKeyEncoding: {
+      type: 'pkcs1',
+      format: 'pem',
+    },
+  })
+
+  const mockConfig = {
+    ssh: {
+      algorithms: {},
+      readyTimeout: 2000,
+      keepaliveInterval: 1000,
+      keepaliveCountMax: 3,
+      term: 'xterm',
+    },
+    user: {
+      privateKey: null,
+    },
+  }
 
   beforeEach(() => {
-    registeredEventHandlers = {}
-
-    mockConfig = {
-      ssh: {
-        algorithms: {
-          kex: ["algo1", "algo2"],
-          cipher: ["cipher1", "cipher2"],
-          serverHostKey: ["ssh-rsa", "ssh-dss"],
-          hmac: ["hmac1", "hmac2"],
-          compress: ["none", "zlib"]
-        },
-        readyTimeout: 20000,
-        keepaliveInterval: 60000,
-        keepaliveCountMax: 10
+    sshServer = new Server(
+      {
+        hostKeys: [privateKey],
       },
-      user: {
-        name: null,
-        password: null,
-        privateKey: null
-      }
-    }
-
-    mockSSH2Client = {
-      on: jest.fn((event, handler) => {
-        registeredEventHandlers[event] = handler
-      }),
-      connect: jest.fn(() => {
-        process.nextTick(() => {
-          // By default, emit ready event unless test modifies this behavior
-          if (registeredEventHandlers.ready) {
-            registeredEventHandlers.ready()
+      (client) => {
+        client.on('authentication', (ctx) => {
+          if (
+            ctx.method === 'password' &&
+            ctx.username === TEST_CREDENTIALS.username &&
+            ctx.password === TEST_CREDENTIALS.password
+          ) {
+            ctx.accept()
+          } else {
+            ctx.reject()
           }
         })
-      }),
-      shell: jest.fn(),
-      end: jest.fn()
-    }
 
-    SSH2.Client.mockImplementation(() => mockSSH2Client)
+        client.on('ready', () => {
+          client.on('session', (accept) => {
+            const session = accept()
+            session.once('pty', (accept) => {
+              accept()
+            })
+            session.once('shell', (accept) => {
+              const stream = accept()
+              stream.write('Connected to test server\r\n')
+            })
+          })
+        })
+      }
+    )
+
+    sshServer.listen(TEST_PORT)
     sshConnection = new SSHConnection(mockConfig)
   })
 
   afterEach(() => {
-    jest.clearAllMocks()
-  })
-
-  describe("connect", () => {
-    it("should handle immediate connection errors", () => {
-      const mockCreds = {
-        host: "localhost",
-        port: 22,
-        username: "user",
-        password: "pass"
-      }
-
-      // Mock the connect method to throw an error immediately
-      mockSSH2Client.connect.mockImplementation(() => {
-        throw new Error("Spooky Error") // Immediate error
-      })
-
-      return sshConnection.connect(mockCreds).catch((error) => {
-        expect(error).toBeInstanceOf(SSHConnectionError)
-        expect(error.message).toBe("Connection failed: Spooky Error")
-      })
-    })
-
-    it("should connect successfully with password", () => {
-      const mockCreds = {
-        host: "localhost",
-        port: 22,
-        username: "user",
-        password: "pass"
-      }
-
-      return sshConnection.connect(mockCreds).then(() => {
-        expect(mockSSH2Client.connect).toHaveBeenCalledWith(
-          expect.objectContaining({
-            host: mockCreds.host,
-            port: mockCreds.port,
-            username: mockCreds.username,
-            password: mockCreds.password,
-            tryKeyboard: true
-          })
-        )
-      })
-    })
-
-    it("should fail after max authentication attempts", () => {
-      const mockCreds = {
-        host: "localhost",
-        port: 22,
-        username: "user",
-        password: "wrongpass"
-      }
-
-      const attempts = DEFAULTS.MAX_AUTH_ATTEMPTS + 1
-      mockSSH2Client.connect.mockImplementation(() => {
-        process.nextTick(() => {
-          registeredEventHandlers.error(new Error("Authentication failed"))
-        })
-      })
-
-      return sshConnection.connect(mockCreds).catch((error) => {
-        expect(error).toBeInstanceOf(SSHConnectionError)
-        expect(error.message).toBe("All authentication methods failed")
-        expect(mockSSH2Client.connect.mock.calls.length).toBe(attempts)
-      })
-    })
-
-    describe("key authentication", () => {
-      const validPrivateKey = `-----BEGIN RSA PRIVATE KEY-----
-MIIEpTestKeyContentHere
------END RSA PRIVATE KEY-----`
-
-      it("should try private key first when both password and key are provided", () => {
-        const mockCreds = {
-          host: "localhost",
-          port: 22,
-          username: "user",
-          password: "pass",
-          privateKey: validPrivateKey
-        }
-
-        return sshConnection.connect(mockCreds).then(() => {
-          expect(mockSSH2Client.connect).toHaveBeenCalledWith(
-            expect.objectContaining({
-              privateKey: validPrivateKey,
-              username: mockCreds.username
-            })
-          )
-        })
-      })
-
-      it("should fall back to password after key authentication failure", () => {
-        const mockCreds = {
-          host: "localhost",
-          port: 22,
-          username: "user",
-          password: "pass",
-          privateKey: validPrivateKey
-        }
-
-        let authAttempts = 0
-        mockSSH2Client.connect
-          .mockImplementationOnce(() => {
-            process.nextTick(() => {
-              authAttempts += 1
-              registeredEventHandlers.error(
-                new Error("Key authentication failed")
-              )
-            })
-          })
-          .mockImplementationOnce(() => {
-            process.nextTick(() => {
-              authAttempts += 1
-              registeredEventHandlers.ready()
-            })
-          })
-
-        return sshConnection.connect(mockCreds).then(() => {
-          expect(authAttempts).toBe(2)
-          expect(mockSSH2Client.connect).toHaveBeenCalledTimes(2)
-          // Verify second attempt used password
-          expect(mockSSH2Client.connect).toHaveBeenLastCalledWith(
-            expect.objectContaining({
-              password: mockCreds.password
-            })
-          )
-        })
-      })
-
-      it("should reject invalid private key format", () => {
-        const mockCreds = {
-          host: "localhost",
-          port: 22,
-          username: "user",
-          privateKey: "invalid-key-format"
-        }
-
-        return sshConnection.connect(mockCreds).catch((error) => {
-          expect(error).toBeInstanceOf(SSHConnectionError)
-          expect(error.message).toBe("Invalid private key format")
-        })
-      })
-    })
-  })
-
-  describe("shell", () => {
-    beforeEach(() => {
-      sshConnection.conn = mockSSH2Client
-    })
-
-    it("should open shell successfully", () => {
-      const mockStream = {
-        on: jest.fn(),
-        stderr: { on: jest.fn() }
-      }
-
-      mockSSH2Client.shell.mockImplementation((options, callback) => {
-        process.nextTick(() => callback(null, mockStream))
-      })
-
-      return sshConnection.shell().then((result) => {
-        expect(result).toBe(mockStream)
-        expect(sshConnection.stream).toBe(mockStream)
-      })
-    })
-
-    it("should handle shell creation errors", () => {
-      mockSSH2Client.shell.mockImplementation((options, callback) => {
-        process.nextTick(() => callback(new Error("Shell error")))
-      })
-
-      return sshConnection.shell().catch((error) => {
-        expect(error.message).toBe("Shell error")
-      })
-    })
-  })
-
-  describe("resizeTerminal", () => {
-    it("should resize terminal if stream exists", () => {
-      const mockStream = {
-        setWindow: jest.fn()
-      }
-      sshConnection.stream = mockStream
-
-      sshConnection.resizeTerminal(80, 24)
-      expect(mockStream.setWindow).toHaveBeenCalledWith(80, 24)
-    })
-
-    it("should do nothing if stream does not exist", () => {
-      sshConnection.stream = null
-      expect(() => sshConnection.resizeTerminal(80, 24)).not.toThrow()
-    })
-  })
-
-  describe("end", () => {
-    it("should close stream and connection", () => {
-      const mockStream = {
-        end: jest.fn()
-      }
-      sshConnection.stream = mockStream
-      sshConnection.conn = mockSSH2Client
-
+    if (sshConnection) {
       sshConnection.end()
+    }
+    return new Promise((resolve) => {
+      sshServer.close(resolve)
+    })
+  })
 
-      expect(mockStream.end).toHaveBeenCalled()
-      expect(mockSSH2Client.end).toHaveBeenCalled()
-      expect(sshConnection.stream).toBeNull()
-      expect(sshConnection.conn).toBeNull()
+  test('should connect with valid credentials', async () => {
+    const credentials = {
+      host: 'localhost',
+      port: TEST_PORT,
+      username: TEST_CREDENTIALS.username,
+      password: TEST_CREDENTIALS.password,
+    }
+
+    const connection = await sshConnection.connect(credentials)
+    assert.ok(connection, 'Connection should be established')
+  })
+
+  test('should reject connection with invalid credentials', async () => {
+    const invalidCredentials = {
+      host: 'localhost',
+      port: TEST_PORT,
+      username: 'wronguser',
+      password: 'wrongpass',
+    }
+
+    try {
+      await sshConnection.connect(invalidCredentials)
+      assert.fail('Connection should have been rejected')
+    } catch (error) {
+      assert.equal(error.name, 'SSHConnectionError')
+      assert.equal(error.message, 'All authentication methods failed')
+    }
+  })
+
+  test('should connect using private key authentication', async () => {
+    const credentials = {
+      host: 'localhost',
+      port: TEST_PORT,
+      username: TEST_CREDENTIALS.username,
+      privateKey: privateKey,
+    }
+
+    // Update server auth handler to accept private key
+    sshServer.removeAllListeners('connection')
+    sshServer.on('connection', (client) => {
+      client.on('authentication', (ctx) => {
+        if (ctx.method === 'publickey' && ctx.username === TEST_CREDENTIALS.username) {
+          ctx.accept()
+        } else {
+          ctx.reject()
+        }
+      })
+
+      client.on('ready', () => {
+        client.on('session', (accept) => {
+          accept()
+        })
+      })
     })
 
-    it("should handle cleanup when no stream or connection exists", () => {
-      sshConnection.stream = null
-      sshConnection.conn = null
+    const connection = await sshConnection.connect(credentials)
+    assert.ok(connection, 'Connection should be established using private key')
+  })
 
-      expect(() => sshConnection.end()).not.toThrow()
+  test('should reject invalid private key format', async () => {
+    const invalidPrivateKey = 'not-a-valid-private-key-format'
+    const credentials = {
+      host: 'localhost',
+      port: TEST_PORT,
+      username: TEST_CREDENTIALS.username,
+      privateKey: invalidPrivateKey,
+    }
+
+    try {
+      await sshConnection.connect(credentials)
+      assert.fail('Connection should have been rejected')
+    } catch (error) {
+      assert.equal(error.name, 'SSHConnectionError')
+      assert.equal(error.message, 'Invalid private key format')
+    }
+  })
+  test('should resize terminal when stream exists', async () => {
+    const credentials = {
+      host: 'localhost',
+      port: TEST_PORT,
+      username: TEST_CREDENTIALS.username,
+      password: TEST_CREDENTIALS.password,
+    }
+
+    // Connect and create shell
+    await sshConnection.connect(credentials)
+    await sshConnection.shell({ term: 'xterm' })
+
+    // Mock the setWindow method on stream
+    let windowResized = false
+    sshConnection.stream.setWindow = (rows, cols) => {
+      windowResized = true
+      assert.equal(rows, 24)
+      assert.equal(cols, 80)
+    }
+
+    // Test resize
+    sshConnection.resizeTerminal(24, 80)
+    assert.ok(windowResized, 'Terminal should be resized')
+  })
+  test('should try private key first when both password and key are provided', async () => {
+    const authAttemptOrder = []
+
+    sshServer.removeAllListeners('connection')
+    sshServer.on('connection', (client) => {
+      client.on('authentication', (ctx) => {
+        authAttemptOrder.push(ctx.method)
+
+        if (ctx.method === 'publickey' && ctx.username === TEST_CREDENTIALS.username) {
+          return ctx.accept()
+        }
+
+        if (
+          ctx.method === 'password' &&
+          ctx.username === TEST_CREDENTIALS.username &&
+          ctx.password === TEST_CREDENTIALS.password
+        ) {
+          return ctx.accept()
+        }
+
+        ctx.reject(['publickey', 'password'])
+      })
+
+      client.on('ready', () => {
+        client.on('session', (accept) => {
+          const session = accept()
+          session.once('pty', (accept) => accept())
+          session.once('shell', (accept) => accept())
+        })
+      })
     })
+
+    const credentials = {
+      host: 'localhost',
+      port: TEST_PORT,
+      username: TEST_CREDENTIALS.username,
+      privateKey: privateKey,
+      password: TEST_CREDENTIALS.password,
+    }
+
+    const connection = await sshConnection.connect(credentials)
+    assert.ok(connection, 'Connection should be established')
+    assert.ok(
+      authAttemptOrder.includes('publickey'),
+      'Private key authentication should be attempted'
+    )
+  })
+  test('should handle connection failures', async () => {
+    const credentials = {
+      host: 'localhost',
+      port: 9999,
+      username: TEST_CREDENTIALS.username,
+      password: TEST_CREDENTIALS.password,
+    }
+
+    try {
+      await sshConnection.connect(credentials)
+      assert.fail('Connection should have failed')
+    } catch (error) {
+      assert.equal(error.name, 'SSHConnectionError')
+      assert.equal(error.message, 'All authentication methods failed')
+    }
+  })
+
+  test('should handle connection timeout', async () => {
+    const credentials = {
+      host: '240.0.0.0',
+      port: TEST_PORT,
+      username: TEST_CREDENTIALS.username,
+      password: TEST_CREDENTIALS.password,
+    }
+
+    try {
+      await sshConnection.connect(credentials)
+      assert.fail('Connection should have timed out')
+    } catch (error) {
+      assert.equal(error.name, 'SSHConnectionError')
+      assert.equal(error.message, 'All authentication methods failed')
+    }
   })
 })
